move to split GeoJSONVT
