+++ src/mbgl/map/annotation.cpp
@@ -4,6 +4,7 @@
 #include <mbgl/map/live_tile.hpp>
 #include <mbgl/map/map_data.hpp>
 #include <mbgl/util/constants.hpp>
+#include <mbgl/util/geojsonvt_convert.hpp>
 #include <mbgl/util/ptr.hpp>
 
 #include <algorithm>
